Update new course flow
