--- app/views/new/duration.html
@@ -1,18 +1,0 @@
-{% extends "layout.html" %}
-{% set title = "Which phase is this course for?" %}
-{% set backLink = paths.back %}
-
-{% block content %}
-  <h1 class="govuk-heading-xl">
-    {{ title }}
-  </h1>
-  <form action="/new/duration" method="post">
-    {{ macros.datefields('start-date', 'Course start date') }}
-
-    <div class="form-group">
-      <input type="submit" class="govuk-button" value="Continue" />
-    </div>
-  </form>
-
-
-{% endblock %}
