--- app/views/new/equivalency-tests.html
@@ -1,75 +1,0 @@
-{% extends "layout.html" %}
-{% set title = "Tests for candidates without GCSE requirements" %}
-{% set isPrimary = data[code + '-phase'] == 'Primary' %}
-{% set backLink = paths.back %}
-
-{% block content %}
-  <!-- examples:
-      1: https://find-postgraduate-teacher-training.education.gov.uk/course/17B/225K
-      2: https://find-postgraduate-teacher-training.education.gov.uk/course/12U/2527
-      3: https://find-postgraduate-teacher-training.education.gov.uk/course/1PJ/236Y
-      9: https://find-postgraduate-teacher-training.education.gov.uk/course/B60/X341
-  -->
-  <h1 class="govuk-heading-xl">
-    {{ title }}
-  </h1>
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-      <form action="{{ paths.next }}" method="post">
-        <p>Some training providers (and accredited bodies) offer tests to candidates who don’t have the minimum GCSEs (or NARIC-recognised equivalents).</p>
-        <p>These identify whether candidates have the required {% if isPrimary %}English, Maths or Science{% else %}English or Maths{% endif %} skills to start a course (they’re sometimes known as ‘equivalency tests’).</p>
-        <h2 class="govuk-heading-m">Will you receive applications from candidates who need to take an {% if isPrimary %}English, Maths or Science{% else %}English or Maths{% endif %} equivalency test?</h2>
-
-        {% set primarySubjects = [
-          {
-            name: 'English',
-            selected: true
-          },
-          {
-            name: 'Mathematics',
-            selected: true
-          },
-          {
-            name: 'Science',
-            selected: true
-          }
-        ] %}
-
-        {% set secondarySubjects = [
-          {
-            name: 'English',
-            selected: true
-          },
-          {
-            name: 'Mathematics',
-            selected: true
-          }
-        ] %}
-
-        {% set moreEquivalency %}
-          <h3 class="govuk-heading-s">Which subjects will you accept equivalency tests in?</h3>
-          {{ macros.checkboxGroup(code + '-equivalency-subjects', primarySubjects if (isPrimary) else secondarySubjects)}}
-        {% endset %}
-
-        {% set notYetFields = [
-          {
-            name: 'Yes (recommended)',
-            more: moreEquivalency
-          },
-          {
-            name: 'No – UCAS will block applications from these candidates'
-          }
-        ] %}
-
-        {{ macros.radioGroup(code + '-equivalency-test', notYetFields)}}
-
-        <hr class="govuk-section-break govuk-section-break--l">
-        <div class="govuk-form-group">
-          <input type="submit" class="govuk-button" value="Continue" />
-        </div>
-      </form>
-    </div>
-  </div>
-
-
-{% endblock %}
