--- app/views/new/main-language.html
@@ -1,45 +1,0 @@
-{% extends "layout.html" %}
-{% set title = "Is one of these languages more important?" %}
-{% set backLink = paths.back %}
-
-{% block content %}
-  <h1 class="govuk-heading-xl">
-    {{ title }}
-  </h1>
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-      <form action="{{ paths.next }}" method="post">
-        {% set languageFields = [
-            {
-              name: 'French'
-            },
-            {
-              name: 'German'
-            },
-            {
-              name: 'Italian'
-            }
-          ]
-        %}
-
-        {% set more %}
-          <p>Which one?</p>
-          {{ macros.radioGroup(code + '-dominant-language', languageFields, 0, false) }}
-        {% endset %}
-
-        {{ macros.radioGroup(code + '-languages', [
-          {
-            name: 'Yes',
-            more: more
-          },
-          {
-            name: 'No, they are equal'
-          }
-        ], 0, false) }}
-        <input type="submit" class="govuk-button" value="Continue" />
-      </form>
-    </div>
-  </div>
-
-
-{% endblock %}
