+++ app/views/new/minimum-gcses.html
@@ -1,50 +1,84 @@
 {% extends "layout.html" %}
-{% set title = "Candidates with pending GCSEs" %}
+{% set title = "GCSE requirements for applicants" %}
 {% set backLink = paths.back %}
 
 {% block content %}
-  <!-- examples:
-      1: https://find-postgraduate-teacher-training.education.gov.uk/course/17B/225K
-      2: https://find-postgraduate-teacher-training.education.gov.uk/course/12U/2527
-      3: https://find-postgraduate-teacher-training.education.gov.uk/course/1PJ/236Y
-      9: https://find-postgraduate-teacher-training.education.gov.uk/course/B60/X341
+  <!-- Examples:
+    1: https://find-postgraduate-teacher-training.education.gov.uk/course/17B/225K
+    2: https://find-postgraduate-teacher-training.education.gov.uk/course/12U/2527
+    3: https://find-postgraduate-teacher-training.education.gov.uk/course/1PJ/236Y
+    9: https://find-postgraduate-teacher-training.education.gov.uk/course/B60/X341
   -->
-  <h1 class="govuk-heading-xl">
-    {{ title }}
-  </h1>
-  <div class="govuk-grid-row">
+  <form action="{{ paths.next }}" method="post" class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-      <form action="{{ paths.next }}" method="post">
-        <p>By law all applicants of a {% if data[code + '-phase'] == 'Primary' %}primary{% else %}secondary{% endif %} course need a GCSE grade C (or grade 4), or an equivalent,&nbsp;in:</p>
-        <ul class="govuk-list govuk-list--bullet">
-          <li>English</li>
-          <li>Mathematics</li>
-          {% if data[code + '-phase'] == 'Primary' %}
-            <li>Science</li>
-          {% endif %}
-        </ul>
+      <h1 class="govuk-heading-xl">{{ title }}</h1>
+      <p class="govuk-body">By law, all applicants need GCSE grade C/grade 4, or an equivalent, in English and maths - plus science for primary courses.</p>
+      <p class="govuk-body">However, they may still apply if:</p>
+      <ul class="govuk-list govuk-list--bullet">
+        <li>they’re currently taking the GCSE</li>
+        <li>they need to take an equivalency test (this proves they have the skills to start training)</li>
+      </ul>
+      <p class="govuk-body">Choose an option below to show how flexible you are towards these applicants.</p>
 
-        <hr class="govuk-section-break govuk-section-break--m">
-        <h2 class="govuk-heading-m">Will you receive applications from candidates with pending GCSEs?</h2>
-        <p>These are candidates who expect to have the qualifications before the beginning of the course. If their application is successful, they’ll usually get a conditional offer.</p>
+      {% set items = [{
+        value: "Must have the GCSE (least flexible)",
+        text: "1. Must have the GCSE (least flexible)",
+        hint: {
+          text: "You won’t receive applications from anyone without the GCSE, or from anyone who needs to take an equivalency test - UCAS will block these applications."
+        }
+      }, {
+        value: "Taking the GCSE",
+        text: "2: Taking the GCSE",
+        hint: {
+          text: "You will consider applicants with a pending GCSE. But UCAS will block applications from anyone who needs to take an equivalency test."
+        }
+      }, {
+        value: "Equivalency test (most flexible)",
+        text: "3: Equivalency test (most flexible)",
+        hint: {
+          text: "You will consider applicants who need to take an equivalency test, as well as those with a pending GCSE."
+        }
+      }] %}
 
-        {% set notYetFields = [
-          {
-            name: 'Yes (recommended)'
-          },
-          {
-            name: 'No – UCAS will block applications from these candidates'
+      {{ govukRadios({
+        idPrefix: code + "-maths-gcse",
+        name: code + "-maths-gcse",
+        fieldset: {
+          legend: {
+            text: "Maths GCSE",
+            classes: "govuk-fieldset__legend--m"
           }
-        ] %}
+        },
+        items: items
+      }) }}
 
-        {{ macros.radioGroup(code + '-no-qualifications-yet', notYetFields)}}
+      {{ govukRadios({
+        idPrefix: code + "-english-gcse",
+        name: code + "-english-gcse",
+        fieldset: {
+          legend: {
+            text: "English GCSE",
+            classes: "govuk-fieldset__legend--m"
+          }
+        },
+        items: items
+      }) }}
 
-        <div class="govuk-form-group">
-          <input type="submit" class="govuk-button" value="Continue" />
-        </div>
-      </form>
+      {{ govukRadios({
+        idPrefix: code + "-science-gcse",
+        name: code + "-science-gcse",
+        fieldset: {
+          legend: {
+            text: "Science GCSE",
+            classes: "govuk-fieldset__legend--m"
+          }
+        },
+        items: items
+      }) if data[code + "-phase"] == "Primary" }}
+
+      {{ govukButton({
+        text: "Continue"
+      }) }}
     </div>
-  </div>
-
-
+  </form>
 {% endblock %}
