+++ app/views/new/phase.html
@@ -1,45 +1,51 @@
 {% extends "layout.html" %}
-{% set title = "What type of new course?" %}
+{% set title = "What type of course?" %}
 {% set backLink = paths.back %}
 
 {% block content %}
-  <h1 class="govuk-heading-xl">
-    {{ title }}
-  </h1>
-  <div class="govuk-grid-row">
+  <form action="{{ paths.next }}" method="post" class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-      <form action="{{ paths.next }}" method="post">
+      <h1 class="govuk-heading-xl">{{ title }}</h1>
 
-        <h2 class="govuk-heading-m remove-top-margin">Level</h2>
-        {% set fields = [
-            {
-              name: 'Primary'
-            },
-            {
-              name: 'Secondary'
-            },
-            {
-              name: 'Further education'
-            }
-          ]
-        %}
+      {{ govukRadios({
+        idPrefix: code + "-phase",
+        name: code + "-phase",
+        fieldset: {
+          legend: {
+            text: "Level",
+            classes: "govuk-fieldset__legend--m"
+          }
+        },
+        items: [{
+          value: "Primary",
+          text: "Primary"
+        }, {
+          value: "Secondary",
+          text: "Secondary"
+        }, {
+          value: "Further education",
+          text: "Further education"
+        }]
+      }) }}
 
-        {{ macros.radioGroup(code + '-phase', fields) }}
+      {{ govukCheckboxes({
+        idPrefix: code + "-sen",
+        name: code + "-sen",
+        fieldset: {
+          legend: {
+            text: "Special educational needs and disability (SEND)",
+            classes: "govuk-fieldset__legend--m"
+          }
+        },
+        items: [{
+          value: "This course has an additional SEND specialism",
+          text: "This course has an additional SEND specialism"
+        }]
+      }) }}
 
-        <h2 class="govuk-heading-m">Special educational needs and disability (SEND)</h2>
-        {% set fields = [
-            {
-              name: 'This course has an additional SEND specialism'
-            }
-          ]
-        %}
-
-        {{ macros.checkboxGroup(code + '-sen', fields, 0, false) }}
-
-        <input type="submit" class="govuk-button" value="Continue" />
-      </form>
+      {{ govukButton({
+        text: "Continue"
+      }) }}
     </div>
-  </div>
-
-
+  </form>
 {% endblock %}
