+++ app/views/new/placement-locations.html
@@ -1,0 +1,26 @@
+{% extends "layout.html" %}
+{% set title = "Which schools can candidates apply to?" %}
+{% set backLink = paths.back %}
+
+{% block content %}
+  <form action="{{ paths.next }}" method="post" class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+      {{ govukCheckboxes({
+        idPrefix: code + "-locations",
+        name: code + "-locations",
+        fieldset: {
+          legend: {
+            text: title,
+            isPageHeading: true,
+            classes: "govuk-fieldset__legend--xl"
+          }
+        },
+        items: locations
+      }) }}
+
+      {{ govukButton({
+        text: "Continue"
+      }) }}
+    </div>
+  </form>
+{% endblock %}
