+++ app/views/new/training-locations.html
@@ -1,36 +1,46 @@
 {% extends "layout.html" %}
-{% set title = "Pick the locations for this course" %}
+{% set title = "Where does academic training take place?" %}
 {% set backLink = paths.back %}
 
 {% block content %}
-  <h1 class="govuk-heading-xl">
-    {{ title }}
-  </h1>
-  <div class="govuk-grid-row">
+  <form action="{{ paths.next }}" method="post" class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-      <form action="{{ paths.next }}" method="post">
-        {% if data[code + '-full-part'] == 'Full time or part time' %}
-          {% set vacancyFields = [
-              {
-                name: 'Full time places'
-              },
-              {
-                name: 'Part time places'
-              }
-            ]
-          %}
+      {{ govukRadios({
+        hint: {
+          text: "Academic training is teaching that takes place outside of the classroom"
+        },
+        idPrefix: code + "-locations",
+        name: code + "-locations",
+        fieldset: {
+          legend: {
+            text: title,
+            isPageHeading: true,
+            classes: "govuk-fieldset__legend--xl"
+          }
+        },
+        items: [{
+          value: "Leeds West Academy",
+          text: "Leeds West Academy",
+          label: {
+            classes: 'govuk-label--s'
+          },
+          hint: {
+            text: "Intake Lane, Leeds, LS13 1DQ"
+          }
+        }, {
+          divider: "or"
+        }, {
+          value: "host",
+          text: "At the school where the candidate is based"
+        }, {
+          value: "remote",
+          text: "Remotely"
+        }]
+      }) }}
 
-          {% set more %}
-            <p>Does this location offer full time places, part time places or both?</p>
-            {{ macros.checkboxGroup(code + '-location-full-part', vacancyFields) }}
-          {% endset %}
-          {% set modifiedLocations = locations|addMoreToFields(more) %}
-          {{ macros.checkboxGroup(code + '-locations', modifiedLocations) }}
-        {% else %}
-          {{ macros.checkboxGroup(code + '-locations', locations) }}
-        {% endif %}
-        <input type="submit" class="govuk-button" value="Continue" />
-      </form>
+      {{ govukButton({
+        text: "Continue"
+      }) }}
     </div>
-  </div>
+  </form>
 {% endblock %}
